Dimnish memory complexity

+++ GPUPatternMining/GPUPatternMining.Contract/CinsNode.h
@@ -8,15 +8,15 @@ class CinsNode
 public:
 	std::vector<std::unique_ptr<CinsNode>> children;
 	CinsNode* parent;
-	unsigned int instanceId; //4,5,12
-	unsigned int type; //A,B,C
+	unsigned short instanceId; //4,5,12
+	unsigned short type; //A,B,C
 
-	unsigned int addChild(unsigned int instanceId, unsigned int type);
-	CinsNode* addChildPtr(unsigned int instanceId, unsigned int type);
-	CinsNode* indexChild(unsigned int instanceId, unsigned int type);
+	unsigned short addChild(unsigned short instanceId, unsigned short type);
+	CinsNode* addChildPtr(unsigned short instanceId, unsigned short type);
+	CinsNode* indexChild(unsigned short instanceId, unsigned short type);
 	std::vector<CinsNode*> getAncestors();
 	std::vector<CinsNode*> getPath();
-	unsigned int getDepth();
+	unsigned short getDepth();
 	CinsNode();
 	~CinsNode();
 };
